pyproject.toml
[project]
name = "vidai"
version = "0.1.0"
description = "FastAPI service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "pymongo>=4.10.0",
    "redis>=5.2.0",
    "temporalio>=1.8.0",
    "sqlalchemy>=2.0.36",
    "pymysql>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
include = ["app", "tests"]
exclude = [".venv", "**/__pycache__"]
reportMissingImports = true
reportMissingModuleSource = true
reportMissingTypeStubs = true
reportInvalidTypeForm = true
reportUntypedFunctionDecorator = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportArgumentType = false
reportIncompatibleMethodOverride = false
reportReturnType = false
